hoàn thiện CRUD bảng sản phẩm, nguyên liệu, tại cột giá của bảng sản phẩm sửa kiểu BigDecimal sang float

+++ DuAn1/src/main/java/viewmodel/QLHDCT.java
@@ -18,7 +18,7 @@ public class QLHDCT {
     private UUID idHD;
     private UUID idSP;
     private String tenSP;
-    private BigDecimal gia;
+    private float gia;
     private int soLuong;
     private int soBanCT;
 
@@ -37,18 +37,18 @@ public class QLHDCT {
     
     
     
-    public BigDecimal getGia() {
+    public float getGia() {
         return gia;
     }
     public float getTien() {
         
         float giaf;
-        giaf= gia.floatValue();
+        giaf= gia;
         float tien= giaf*soLuong;
         return tien;
     }
 
-    public void setGia(BigDecimal gia) {
+    public void setGia(float gia) {
         this.gia = gia;
     }
     public QLHDCT(HDCT hdct) {
@@ -64,7 +64,7 @@ public class QLHDCT {
     public String getTenSP() {
         return tenSP;
     }
-    public QLHDCT(UUID idHDCT, UUID idHD, UUID idSP, String tenSP, BigDecimal gia, int soLuong, int soBanCT) {
+    public QLHDCT(UUID idHDCT, UUID idHD, UUID idSP, String tenSP, float gia, int soLuong, int soBanCT) {
         this.idHDCT = idHDCT;
         this.idHD = idHD;
         this.idSP = idSP;
